requirements.txt
streamlit>=1.27.0
pandas>=1.5.3
plotly>=5.14.0
psycopg2-binary>=2.9.6
python-dotenv>=1.0.0
openpyxl>=3.1.2
matplotlib>=3.7.2
numpy>=1.24.0
sqlalchemy>=2.0.0
